+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 -   true_z property for components which gets the z_index of the `Component` offset by its parent `GameObject`
 -   `Time.scheduled_call()` for a self-correcting recurring method call on a timed interval.
 -   clone functions for `Group`, `GameObject`, and `Scene`.
+-   `Draw.image` and `Draw.immediate_image` to draw SDL surfaces and textures to the renderer
 
 ### Changed